+++ src/OpenTelemetry.AutoInstrumentation/.publicApi/net462/PublicAPI.Shipped.txt
@@ -31,10 +31,6 @@ OpenTelemetry.AutoInstrumentation.Instrumentations.Log4Net.TraceContextInjection
 OpenTelemetry.AutoInstrumentation.Instrumentations.MongoDB.Integrations.MongoClientIntegrationExecute
 OpenTelemetry.AutoInstrumentation.Instrumentations.MongoDB.Integrations.MongoClientIntegrationExecuteAsync
 OpenTelemetry.AutoInstrumentation.Instrumentations.NServiceBus.EndpointConfigurationIntegration
-OpenTelemetry.AutoInstrumentation.Instrumentations.RabbitMq6.Integrations.AsyncDefaultBasicConsumerIntegration
-OpenTelemetry.AutoInstrumentation.Instrumentations.RabbitMq6.Integrations.DefaultBasicConsumerIntegration
-OpenTelemetry.AutoInstrumentation.Instrumentations.RabbitMq6.Integrations.ModelBaseBasicGetIntegration
-OpenTelemetry.AutoInstrumentation.Instrumentations.RabbitMq6.Integrations.ModelBasicPublishIntegration
 OpenTelemetry.AutoInstrumentation.Instrumentations.Wcf.Client.WcfClientIntegration
 OpenTelemetry.AutoInstrumentation.Instrumentations.Wcf.Service.ServiceHostIntegration
 override OpenTelemetry.AutoInstrumentation.CallTarget.CallTargetReturn<T>.ToString() -> string!
