+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/RabbitMqLegacy/IntegrationConstants.cs
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-namespace OpenTelemetry.AutoInstrumentation.Instrumentations.RabbitMq6;
+namespace OpenTelemetry.AutoInstrumentation.Instrumentations.RabbitMqLegacy;
 
 internal static class IntegrationConstants
 {
